+++ events-app/tickets/events-service/src/main/java/com/prk/EventsServiceApplication.java
@@ -1,0 +1,11 @@
+package com.prk;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class EventsServiceApplication {
+    public static void main(String[] args) {
+        SpringApplication.run(EventsServiceApplication.class, args);
+    }
+}
