+++ events-app/tickets/events-service/src/main/java/com/prk/events/ControllerExceptionHandler.java
@@ -1,0 +1,17 @@
+package com.prk.events;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.web.ErrorResponse;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.bind.annotation.RestControllerAdvice;
+
+import java.util.NoSuchElementException;
+
+@RestControllerAdvice
+public class ControllerExceptionHandler {
+
+    @ExceptionHandler(NoSuchElementException.class)
+    public ErrorResponse notFound(NoSuchElementException ex) {
+        return ErrorResponse.create(ex, HttpStatus.NOT_FOUND, ex.getMessage());
+    }
+}
