+++ events-app/tickets/events-service/src/main/java/com/prk/events/EventRepository.java
@@ -1,0 +1,9 @@
+package com.prk.events;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import java.util.List;
+
+public interface EventRepository extends JpaRepository<Event, Integer> {
+
+    List<Event> findByOrganizerId(int organizerId);
+}
