+++ events-app/tickets/events-service/src/main/java/com/prk/events/ProductRepository.java
@@ -1,0 +1,9 @@
+package com.prk.events;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import java.util.List;
+
+public interface ProductRepository extends JpaRepository<Product, Integer> {
+
+    List<Product> findByEventId(int eventId);
+}
