+++ events-app/tickets/events-service/src/main/java/com/prk/events/VenueRepository.java
@@ -1,0 +1,6 @@
+package com.prk.events;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface VenueRepository extends JpaRepository<Venue, Integer> {
+}
