+++ events-app/tickets/events-service/src/main/resources/application.yml
@@ -1,0 +1,14 @@
+server:
+  port: 8081
+spring:
+  datasource:
+    url:  jdbc:postgresql://localhost:5432/tickets
+    username: tickets
+    password: tickets
+  jpa:
+    hibernate:
+      ddl-auto: none
+    open-in-view: false
+  sql:
+    init:
+      mode: never
