+++ events-app/tickets/events-service/src/main/resources/schema.sql
