+++ events-app/tickets/registration-service/src/main/java/com/prk/RegistrationServiceApplication.java
@@ -1,0 +1,11 @@
+package com.prk;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class RegistrationServiceApplication {
+    public static void main(String[] args) {
+        SpringApplication.run(RegistrationServiceApplication.class, args);
+    }
+}
