+++ events-app/tickets/registration-service/src/main/java/com/prk/registration/RegistrationController.java
@@ -1,0 +1,53 @@
+package com.prk.registration;
+
+import jakarta.validation.Valid;
+import org.springframework.web.bind.annotation.*;
+
+import java.util.NoSuchElementException;
+import java.util.UUID;
+
+@RestController
+@RequestMapping(path = "/registrations")
+public class RegistrationController {
+
+    private final RegistrationRepository registrationRepository;
+
+    public RegistrationController(RegistrationRepository registrationRepository) {
+        this.registrationRepository = registrationRepository;
+    }
+
+    @PostMapping
+    public Registration create(@RequestBody @Valid Registration registration) {
+        String ticketCode = UUID.randomUUID().toString();
+        return registrationRepository.save(new Registration(
+                null,   // supplied by mongo
+                registration.productId(),
+                ticketCode,
+                registration.attendeeName()));
+    }
+
+    @GetMapping(path = "/{ticketCode}")
+    public Registration get(@PathVariable("ticketCode") String ticketCode) {
+        return registrationRepository.findByTicketCode(ticketCode)
+                .orElseThrow(() -> new NoSuchElementException("registration by ticket code " + ticketCode + " not found"));
+    }
+
+    @PutMapping
+    public Registration update(@RequestBody Registration registration) {
+        // lookup existing registration by ticket-code
+        String ticketCode = registration.ticketCode();
+        var existing = registrationRepository.findByTicketCode(ticketCode)
+                .orElseThrow(() -> new NoSuchElementException("registration with ticket code " + ticketCode + " not found"));
+        // save, only update attendee name
+        return registrationRepository.save(new Registration(
+                existing.id(),
+                existing.productId(),
+                ticketCode,
+                registration.attendeeName()));
+    }
+
+    @DeleteMapping(path = "/{ticketCode}")
+    public void delete(@PathVariable("ticketCode") String ticketCode) {
+        registrationRepository.deleteByTicketCode(ticketCode);
+    }
+}
