+++ events-app/tickets/registration-service/src/main/java/com/prk/registration/RegistrationRepository.java
@@ -1,0 +1,11 @@
+package com.prk.registration;
+
+import org.springframework.data.mongodb.repository.MongoRepository;
+import java.util.Optional;
+
+public interface RegistrationRepository extends MongoRepository<Registration, String> {
+
+   Optional<Registration> findByTicketCode(String ticketCode);
+
+   void deleteByTicketCode(String ticketCode);
+}
