+++ events-app/tickets/registration-service/src/main/resources/application.yaml
@@ -1,0 +1,7 @@
+tickets:
+  events:
+    url: http://localhost:8081
+spring:
+  data:
+    mongodb:
+      uri: mongodb://root:example@localhost:27017/tickets?authSource=admin
