--- events-app/tickets/src/main/java/com/prk/tickets/ControllerExceptionHandler.java
@@ -1,17 +1,0 @@
-package com.prk.tickets;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.web.ErrorResponse;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.RestControllerAdvice;
-
-import java.util.NoSuchElementException;
-
-@RestControllerAdvice
-public class ControllerExceptionHandler {
-
-    @ExceptionHandler(NoSuchElementException.class)
-    public ErrorResponse notFound(NoSuchElementException ex) {
-        return ErrorResponse.create(ex, HttpStatus.NOT_FOUND, ex.getMessage());
-    }
-}
