--- events-app/tickets/src/main/java/com/prk/tickets/TicketsApplication.java
@@ -1,13 +1,0 @@
-package com.prk.tickets;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class TicketsApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(TicketsApplication.class, args);
-	}
-
-}
