--- events-app/tickets/src/main/java/com/prk/tickets/events/Event.java
@@ -1,78 +1,0 @@
-package com.prk.tickets.events;
-
-import java.time.LocalDate;
-import jakarta.persistence.*;
-
-@Entity
-@Table(name = "events")
-public class Event {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
-
-    @Column(nullable = false)
-    private String name;
-
-    @ManyToOne(optional = false)
-    @JoinColumn(name = "organizer_id", referencedColumnName = "id", nullable = false)
-    private Organizer organizer;
-
-    @ManyToOne(optional = false)
-    @JoinColumn(name = "venue_id", referencedColumnName = "id", nullable = false)
-    private Venue venue;
-
-    @Column
-    private LocalDate startDate;
-
-    @Column
-    private LocalDate endDate;
-
-    public int getId() {
-        return id;
-    }
-
-    public void setId(int id) {
-        this.id = id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public Organizer getOrganizer() {
-        return organizer;
-    }
-
-    public void setOrganizer(Organizer organizer) {
-        this.organizer = organizer;
-    }
-
-    public Venue getVenue() {
-        return venue;
-    }
-
-    public void setVenue(Venue venue) {
-        this.venue = venue;
-    }
-
-    public LocalDate getStartDate() {
-        return startDate;
-    }
-
-    public void setStartDate(LocalDate startDate) {
-        this.startDate = startDate;
-    }
-
-    public LocalDate getEndDate() {
-        return endDate;
-    }
-
-    public void setEndDate(LocalDate endDate) {
-        this.endDate = endDate;
-    }
-}
