--- events-app/tickets/src/main/java/com/prk/tickets/events/EventRepository.java
@@ -1,9 +1,0 @@
-package com.prk.tickets.events;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import java.util.List;
-
-public interface EventRepository extends JpaRepository<Event, Integer> {
-
-    List<Event> findByOrganizerId(int organizerId);
-}
