--- events-app/tickets/src/main/java/com/prk/tickets/events/Organizer.java
@@ -1,44 +1,0 @@
-package com.prk.tickets.events;
-
-import jakarta.persistence.*;
-
-@Entity
-@Table(name = "organizers")
-public class Organizer {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
-
-    @Column(nullable = false)
-    private String name;
-
-    @Column(nullable = false)
-    private String description;
-
-    public int getId() {
-        return id;
-    }
-
-    public void setId(int id) {
-        this.id = id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public String getDescription() {
-        return description;
-    }
-
-    public void setDescription(String description) {
-        this.description = description;
-    }
-}
-
-
