--- events-app/tickets/src/main/java/com/prk/tickets/events/OrganizerRepository.java
@@ -1,6 +1,0 @@
-package com.prk.tickets.events;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface OrganizerRepository extends JpaRepository<Organizer, Integer> {
-}
