--- events-app/tickets/src/main/java/com/prk/tickets/events/ProductRepository.java
@@ -1,9 +1,0 @@
-package com.prk.tickets.events;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import java.util.List;
-
-public interface ProductRepository extends JpaRepository<Product, Integer> {
-
-    List<Product> findByEventId(int eventId);
-}
