--- events-app/tickets/src/main/java/com/prk/tickets/events/Venue.java
@@ -1,64 +1,0 @@
-package com.prk.tickets.events;
-
-import jakarta.persistence.*;
-
-@Entity
-@Table(name = "venues")
-public class Venue {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
-
-    @Column(nullable = false)
-    private String name;
-
-    @Column
-    private String street;
-
-    @Column
-    private String city;
-
-    @Column
-    private String country;
-
-    public int getId() {
-        return id;
-    }
-
-    public void setId(int id) {
-        this.id = id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public String getStreet() {
-        return street;
-    }
-
-    public void setStreet(String street) {
-        this.street = street;
-    }
-
-    public String getCity() {
-        return city;
-    }
-
-    public void setCity(String city) {
-        this.city = city;
-    }
-
-    public String getCountry() {
-        return country;
-    }
-
-    public void setCountry(String country) {
-        this.country = country;
-    }
-}
