--- events-app/tickets/src/main/java/com/prk/tickets/events/VenueRepository.java
@@ -1,6 +1,0 @@
-package com.prk.tickets.events;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface VenueRepository extends JpaRepository<Venue, Integer> {
-}
