--- events-app/tickets/src/main/java/com/prk/tickets/registration/RegistrationRepository.java
@@ -1,11 +1,0 @@
-package com.prk.tickets.registration;
-
-import org.springframework.data.mongodb.repository.MongoRepository;
-import java.util.Optional;
-
-public interface RegistrationRepository extends MongoRepository<Registration, String> {
-
-   Optional<Registration> findByTicketCode(String ticketCode);
-
-   void deleteByTicketCode(String ticketCode);
-}
