--- events-app/tickets/src/main/resources/application.properties
@@ -1,2 +1,0 @@
-spring.application.name=tickets
-spring.data.mongodb.uri=mongodb://root:example@localhost:27017/tickets?authSource=admin
