--- events-app/tickets/src/test/java/com/prk/tickets/TicketsApplicationTests.java
@@ -1,13 +1,0 @@
-package com.prk.tickets;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class TicketsApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
